requirements.txt
PyYAML >= 3.0
six >= 1.10
